llama: take the float out of the forward pass, 10x
The first port quantized the weights and kept llama2.c's fp32
activations: quantize, int8 matmul, dequantize to float, RMSNorm and
SwiGLU and softmax in float, quantize again. So int8 covered the
O(weights) work and float covered the O(dim) work around it, which on a
part with no FPU is backwards -- 259,328 integer MACs at 18 cycles,
wrapped in about 24,000 float operations at 250. Soft float was 80% of
the profile and the model's own arithmetic was 8%.

Now nothing in the forward pass is floating point, and neither is
anything in the weight file. 118 ms a token against 1214, and 131
against 1817 at 32 tokens -- the attention path no longer costs float
per position, so context barely moves it. matmul is 59% of the profile
where it was 8%, which is the shape it should have had.

Every activation vector carries a power-of-two exponent and nothing
else, so exponents add and no scale is ever multiplied into data. That
works because a non-power-of-two scale appears in exactly one place,
RMSNorm's sqrt(n)/sqrt(sum x^2), and is folded into the data there.
RMSNorm is also the one operation easier in fixed point than in float:
the input exponent divides out, which is what normalising means.

Two rules keep every product inside an int32. Nothing widens, because
gcc compiles (int64_t)a * b to a call to __muldi3 and that was 5% of the
old profile by itself; operands are cut to fifteen bits first. And
nothing divides that need not: softmax normalises with one reciprocal a
head rather than one a score, and attention weights come out in Q7 so
the weighted sum of values carries a shift instead of a division.

The KV cache is int8 with one exponent per (layer, position), a quarter
of what fp32 took. Positions therefore disagree about exponent, so
scores and values are brought to the smallest one in the cache before
being compared or summed; shifting down is exact and shifting up would
overflow.

Accuracy is measured, not assumed. The new test feeds a fixed token
sequence to this and to an fp32 numpy reference and compares the top-1
prediction at every position: 45 of 45. Free-running generation follows
llama2.c's fp32 output for 42 tokens and then takes a different but
equally sensible turn, which is what one differing logit does under
argmax and says nothing about the arithmetic -- which is why the old
exact-match test was replaced rather than relaxed.

+++ llama/runner.c
@@ -33,7 +33,7 @@ int llama_run(llama_model *m, llama_tokenizer *tok,
 	token = prompt_tokens[0];
 	for (pos = 0; pos < steps; pos++) {
 		uint32_t t0 = llama_now_us();
-		float *logits = llama_forward(m, token, pos);
+		int32_t *logits = llama_forward(m, token, pos);
 
 		stats->forward_us += llama_now_us() - t0;
 		stats->macs += llama_macs(m, pos);
@@ -44,7 +44,7 @@ int llama_run(llama_model *m, llama_tokenizer *tok,
 		} else {
 			/* Argmax over the logits. */
 			int best = 0, i;
-			float best_val = logits[0];
+			int32_t best_val = logits[0];
 
 			for (i = 1; i < m->cfg.vocab_size; i++)
 				if (logits[i] > best_val) {
